requirements.txt
asgiref==3.9.1
decouple==0.0.7
Django==5.2.4
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
uritemplate==4.2.0
